backend/app/requirements.txt
fastapi
uvicorn
python-dotenv
openai 
torch
transformers
huggingface_hub
Pillow

